test(test-optimization): split Vitest TIA tests
Move the TIA integration cases to their own matrix shard so test execution and coverage conversion can run in parallel with the core Vitest cases.
